v2/login/web/requirements.txt
flask
pymongo
flask-cors
Flask-And-Redis
flask-restful
